refactored Job Engine module
