+++ assembly/broker/configurations/locator.xml
@@ -19,13 +19,7 @@
         <api>org.eclipse.kapua.service.certificate.CertificateFactory</api>
         <api>org.eclipse.kapua.service.certificate.info.CertificateInfoService</api>
         <api>org.eclipse.kapua.service.certificate.info.CertificateInfoFactory</api>
-
-        <!-- TODO EXT-CAMEL only for test remove when jobs will be defined in their own container -->
-        <!-- job services -->
-        <api>org.eclipse.kapua.job.engine.JobEngineService</api>
-        <api>org.eclipse.kapua.job.engine.JobEngineFactory</api>
-        <api>org.eclipse.kapua.job.engine.queue.QueuedJobExecutionService</api>
-        <api>org.eclipse.kapua.job.engine.queue.QueuedJobExecutionFactory</api>
+        
         <provide>
             <interceptor>annotation</interceptor>
             <with>Impl</with>
