+++ job-engine/jbatch/src/main/java/org/eclipse/kapua/job/engine/jbatch/JobengineJbatchModule.java
@@ -10,24 +10,16 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.client;
+package org.eclipse.kapua.job.engine.jbatch;
 
+import org.eclipse.kapua.commons.core.AbstractKapuaModule;
 import org.eclipse.kapua.job.engine.JobEngineFactory;
-import org.eclipse.kapua.job.engine.JobStartOptions;
+import org.eclipse.kapua.job.engine.JobEngineService;
 
-import javax.inject.Singleton;
-
-/**
- * {@link JobEngineFactory} remote client implementation
- *
- * @since 1.5.0
- */
-@Singleton
-public class JobEngineFactoryClient implements JobEngineFactory {
-
+public class JobengineJbatchModule extends AbstractKapuaModule {
     @Override
-    public JobStartOptions newJobStartOptions() {
-        return new JobStartOptionsClient();
+    protected void configureModule() {
+        bind(JobEngineFactory.class).to(JobEngineFactoryJbatch.class);
+        bind(JobEngineService.class).to(JobEngineServiceJbatch.class);
     }
-
 }
